In WorkerTest, just run worker in main test thread
I'm not sure why I originally wrote this to spawn a separate thread
since we can just rely on the fact that pushing a `:shutdown` message
will make the `Worker#run` exit. And the whole "worker runs in a
separate thread" thing is really just an implementation detail of
`AsyncBatchProcessor`, not `Worker`.