Remove minitest-ci
This gem is useful for CircleCI, where there's a dedicated tab to show
details about failed tests, which is powered by the XML test reports
generated by minitest-ci. Now that we've switched to GitHub Actions,
there's no special processing of the test report files. We could save
the raw XML files as artifacts, but now that we're running tests in
verbose mode, they would only add a bit more granularity to the info
we already display in the CI log. So, let's just remove minitest-ci
for now - if we miss it, we can always add it back!