Collect code coverage when running tests on CI
The coverage report will be uploaded as an artifact for anyone to
inspect. Artifact storage expires after 90 days, which should be plenty
of time, but we could switch from the HTML reports to the console
reporter if we decide we want the results to persist forever. (The
console reporter just shows line numbers in the output, though, whereas
the HTML reports display the actual code with each line highlighted
according to its coverage.)